DotaCoach/backend/services/image_analyzer.py
# backend/services/image_analyzer.py
from typing import Dict, List, Tuple
from PIL import Image
import numpy as np

class ImageAnalyzer:
    def __init__(self):
        # In production, this would load a trained model
        self.hero_templates = self._load_hero_templates()
        self.draft_positions = self._load_draft_positions()

    def _load_hero_templates(self) -> Dict[int, np.ndarray]:
        # Mock template data - in production would load actual hero portraits
        return {
            1: np.zeros((64, 64, 3)),  # Mock template for Anti-Mage
            2: np.zeros((64, 64, 3)),  # Mock template for Axe
            # Add more hero templates...
        }

    def _load_draft_positions(self) -> Dict[str, List[Tuple[int, int, int, int]]]:
        # Define bounding boxes for hero positions in draft screen
        return {
            "allies": [
                (100, 100, 164, 164),  # Position 1
                (200, 100, 264, 164),  # Position 2
                # Add more positions...
            ],
            "enemies": [
                (500, 100, 564, 164),  # Position 1
                (600, 100, 664, 164),  # Position 2
                # Add more positions...
            ]
        }

    def preprocess_image(self, image: Image.Image) -> np.ndarray:
        # Convert PIL Image to numpy array and normalize
        img_array = np.array(image)
        img_array = img_array / 255.0
        return img_array

    def detect_heroes(self, image: Image.Image) -> Dict[str, List[int]]:
        """
        Detect heroes in the draft screen image
        Returns dict with lists of hero IDs for allies and enemies
        """
        # In production, this would use actual computer vision
        # For prototype, return mock data
        return {
            "allies": [1, 2],
            "enemies": [3, 4]
        }

    def determine_draft_stage(self, detected_heroes: Dict[str, List[int]]) -> int:
        """Calculate the current draft stage based on picked heroes"""
        total_picks = len(detected_heroes["allies"]) + len(detected_heroes["enemies"])
        return total_picks + 1

    async def analyze_draft_image(self, image: Image.Image) -> Tuple[Dict[str, List[int]], int, float]:
        """
        Analyze draft image and return detected heroes, draft stage, and confidence
        """
        try:
            # Preprocess image
            processed_image = self.preprocess_image(image)
            
            # Detect heroes
            detected_heroes = self.detect_heroes(image)
            
            # Determine draft stage
            draft_stage = self.determine_draft_stage(detected_heroes)
            
            # Mock confidence score - in production would be actual model confidence
            confidence_score = 0.85
            
            return detected_heroes, draft_stage, confidence_score
            
        except Exception as e:
            raise ValueError(f"Error analyzing draft image: {str(e)}")