DotaCoach/backend/requirements.txt
# requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
httpx==0.26.0
pydantic==2.6.1
python-jose[cryptography]==3.3.0
pillow==10.2.0
numpy==1.26.3